+++ src/callbacks/postPEffectUpdate.ts
@@ -1,6 +1,5 @@
 import { isChildPlayer } from "isaacscript-common";
 import * as flipCustom from "../features/items/flipCustom";
-import * as sawblade from "../features/items/sawblade";
 import { fastTravelPostPEffectUpdate } from "../features/optional/major/fastTravel/callbacks/postPEffectUpdate";
 import * as startWithD6 from "../features/optional/major/startWithD6";
 import * as roll from "../features/optional/other/roll";
@@ -24,7 +23,6 @@ export function main(player: EntityPlayer): void {
 
   // Items
   flipCustom.postPEffectUpdate(player);
-  sawblade.postPEffectUpdate(player);
 
   // Other
   roll.postPEffectUpdate(player);